whisperFastAPICelery-proyecto-tfg/requirements.txt
celery==5.2.7
deep-translator==1.10.1
fastapi==0.95.2
Jinja2==3.1.2
openai-whisper==20230124
python-multipart==0.0.5
redis==4.5.4
torch==2.0.1
tzdata==2023.3
uvicorn==0.22.0
yt_dlp==2023.3.4